* TO DO  * x. TRY X11 HEADLESS SETUP  * x. BLIT THE MULTI SAMPLE BUFFER TO A REFULAR ONE AND READ FROM THERE  * x. Anti aliasing for RGB images  *  . Depth image -> more precise from 255  * x. Compress the mask images  *  . save the results to a new folder under ./build  --> experiments/build/results  *  . skip the folders already existed

+++ experiment/Includes/Color.h
@@ -9,7 +9,7 @@ public:
 	Color();
 
     static int toUniqueInt( const int & r, const int & g, const int & b );
-    static color3 toUniqueColor ( const size_t & i);
+    static color3 toUniqueColor ( const int & i);
 
 
     const static color4 BLACK ;
